Refactor encode/round
